+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Generalize data aggregation within analytics folder
 - Updated README release steps
 - Added local S3 docker container via localstack
+- Added error handling to citation API
 
 ## Fixed
 - Changed HATHI_DATAFILES outdated link in development, example, and local yaml files